requirements.txt
# Production dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
Pillow>=9.5.0
jinja2>=3.1.2
python-dotenv>=1.0.0
email-validator>=2.1.0
pydantic>=1.10.0,<2.0.0
httpx>=0.24.0
streamlit>=1.24.0
numpy==1.24.3
pandas==2.0.3

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
mypy>=1.0.0
python-docx>=1.1.0 # For DOCX generation